backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
anthropic==0.7.7
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
tenacity==8.2.3
pandas==2.1.3
numpy==1.26.2
redis==5.0.1
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0